+++ www/go/core/model/Search.php
@@ -98,11 +98,11 @@ class Search extends AclOwnerEntity {
 
 	protected static function defineMapping() {
 		return parent::defineMapping()
-										->addTable('core_search', 's')
+										->addTable('core_search', 'search')
 										->setQuery(
 														(new Query())
 														->select("e.clientName AS entity")
-														->join('core_entity', 'e', 'e.id = s.entityTypeId')
+														->join('core_entity', 'e', 'e.id = search.entityTypeId')
 		);
 	}
 
@@ -113,7 +113,7 @@ class Search extends AclOwnerEntity {
 	 * @param int $level
 	 */
 	public static function applyAclToQuery(Query $query, $level = Acl::LEVEL_READ, $userId = null, $groups = null) {
-		Acl::applyToQuery($query, 's.aclId', $level, $userId, $groups);
+		Acl::applyToQuery($query, 'search.aclId', $level, $userId, $groups);
 		
 		return $query;
 	}
@@ -122,7 +122,7 @@ class Search extends AclOwnerEntity {
 		return parent::defineFilters()
 						->add("link", function(Criteria $criteria, $value, Query $query) {
 							
-							$on = 's.entityId = link.toId AND s.entityTypeId = link . toEntityTypeId';							
+							$on = 'search.entityId = link.toId AND search.entityTypeId = link . toEntityTypeId';
 								
 							$query->join('core_link', 'link', $on); 
 
@@ -156,70 +156,42 @@ class Search extends AclOwnerEntity {
 							
 						})
 						->add('text', function(Criteria $criteria, $value, Query $query) {
-
-//							$criteria->andWhere(
-//								(new Criteria())
-//								->orWhere('keywords','like', '%' . preg_replace('/\\s/', '%', $value) . '%')
-//							);
-
-							$i = 0;
-
-							$words = SearchableTrait::splitTextKeywords($value);
-
-							$words = array_unique($words);
-
-							foreach($words as $word) {
-								$query->join("core_search_word", 'w'.$i, 'w'.$i.'.searchId = s.id');
-								$criteria->where('w'.$i.'.word', 'LIKE', $word . '%');
-								$i++;
-							}
-
-//							 $value = static::convertQuery($value);
-//
-//							 $criteria->where('MATCH (s.keywords) AGAINST (:keyword1 IN BOOLEAN MODE)')
-//							 ->bind(':keyword1', $value)
-//							 ->bind(':keyword2', $value);
-//
-//							 // Order by relevance
-//							 $query->orderBy([new Expression('MATCH (s.keywords) AGAINST (:keyword2 IN BOOLEAN MODE) DESC')]);
+							SearchableTrait::addCriteria( $criteria, $query, $value);
 						});					
 	}
 
 	public static function sort(\go\core\orm\Query $query, array $sort)
 	{
-		if(empty($sort)) {
-			$sort['s.id'] = 'DESC'; //sort by id is fast. Sorting by modified at is slow
-		}
-
-		return parent::sort($query, $sort);
+		//no sorting. Too big tables!
+		return $query;
 	}
 
-	 public static function convertQuery($value) {
-
-	 		//first occuring quote type will be used for tokenizing.
-	 		$doublepos = strpos($value, '"');
-	 		$singlepos = strpos($value, "'");
-	 		$quote = '"';
-	 		if($singlepos !== false && $singlepos > $doublepos) {
-	 			$quote = "'";
-	 		}
-
-	 		//https://stackoverflow.com/questions/2202435/php-explode-the-string-but-treat-words-in-quotes-as-a-single-word
-	 		preg_match_all('/'.$quote.'(?:\\\\.|[^\\\\'.$quote.'])*'.$quote.'|\S+/', $value, $tokens);
-
-	 		$str = "";
-
-	 		foreach($tokens[0] as $token) {
-					
-	 				if(substr($token, -1,1) !== $quote) {
-	 					$token = $token .= '*';
-	 				}
-	 				$str .= '+' . $token . ' ';
-	 		}
-
-	 		return $str;
-	 }
-	
+//	 public static function convertQuery($value) {
+//
+//	 		//first occuring quote type will be used for tokenizing.
+//	 		$doublepos = strpos($value, '"');
+//	 		$singlepos = strpos($value, "'");
+//	 		$quote = '"';
+//	 		if($singlepos !== false && $singlepos > $doublepos) {
+//	 			$quote = "'";
+//	 		}
+//
+//	 		//https://stackoverflow.com/questions/2202435/php-explode-the-string-but-treat-words-in-quotes-as-a-single-word
+//	 		preg_match_all('/'.$quote.'(?:\\\\.|[^\\\\'.$quote.'])*'.$quote.'|\S+/', $value, $tokens);
+//
+//	 		$str = "";
+//
+//	 		foreach($tokens[0] as $token) {
+//
+//	 				if(substr($token, -1,1) !== $quote) {
+//	 					$token = $token .= '*';
+//	 				}
+//	 				$str .= '+' . $token . ' ';
+//	 		}
+//
+//	 		return $str;
+//	 }
+//
 	
 	// protected static function textFilterColumns() {
 	// 	return ['keywords'];